backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy>=2.0
passlib[bcrypt]
python-jose[cryptography]
python-multipart
